requirements.txt
pipreqs
numpy==2.1.1
opencv_python==4.10.0.84
pdf2image==1.17.0
Pillow==10.4.0
pytesseract==0.3.13
